Add Lodestone character sync & all tracking models
Allows a user to associate a Lodestone character with their account.
When new users are discovered, their public collection data is stored in
the database. Adds all tracking models necessary to store this
collection data.

TODO: Allow characters to be selected by offline users. Allow users to
verify ownership of their Lodestone character. Schedule a task to
regularly sync user-characters with the Lodestone.