Add character privacy settings
Allows verified users to make their characters' collections private.
Private collections cannot be browsed by other users, and later will be
excluded from the API. If a user is tracking a character that becomes
private, that tracking will be halted and the user will need to select a
new character to track.