Add more character security
Ensure that users are not able to do stuff with characters that do not
exist. Ensure that users are not able to verify characters they do not
have selected. Ensure that users are not able to verify characters that
are already verified.

Also adds a proper 404 to /achievements/types for web crawlers.