Improve achievements handling
When a character is fetched, use the achievement count to keep track of
whether or not their achievements are set to private. Move the privacy
alert to the actual achievements pages. Properly clarify if the
character's achievements are missing or private. Updates the character
refresh to always make a fresh call to the API.