Add user & character API endpoints
Allows users to search for a user by their Discord ID or a character by
their Lodestone character ID. For users, it will retrieve all of that
user's verified, public characters. Character data includes the standard
Lodestone data and all collection data (including Triple Triad if the
user has a collection.)

Refactors `triple_triad` to be a user method, which the character will
call for its verified user if present.