Source improvements, TC/Mogboard integration
After collectables are loaded, retrieve their related item IDs which can
be used to link them to other sites for crafting/pricing/etc. Adds these
links to the existing tables. Achievement items link to the new TC
database now instead of Garland Tools, for consistency.

Add relevant IDs to sources to link them to achievements and recipes.
Achievement and Crafting sources are now created automatically using
this new system, and so they are removed from the extracts.

Adds commands to Emotes to help link them to their items, and displays
this to the user since it's useful information.