Add filters to most collectables
Adds the option to filter collectable data for everything except
Achievements and Orchestrions. The filters consume & expand the existing
"Hide Owned" option and also add the ability to completely exclude
premium sources. This will later be expanded to include an option for
limited-time sources. Adds a new Collectable concern which is used to
share logic among collectables, including the new filtration.