Add Leaderboards
Allows users to see the top 10 characters for each collectable type.
This can be filtered to a specific data center or world. The currently
selected character will also be displayed relative to these rankings,
even if they are on another world/data center.

Adds a new counter for achievement points in order to facilitate this
category on the leaderboards. Adds indexes to all character counts.