Update event quest sources
If a collectable comes from an event quest, mark the source type as
Event instead of Quest so it can be filtered properly. Update the
related links to also link to quests for event sources where
appropriate. Removes redundant event quest sources. Fixes the Quest
load task to update quests if necessary.