Try to create quest relations for event sources
When saving an event source, look for a quest with a name matching the
source text. If found, add the relation to that quest.

Updates the relation assignment to remove the relation information if
it is not applicable to the current source type, useful if the type
changes to something that wouldn't have a relation.