Update character fetching for upcoming API changes
Switches to a locally installed gem on the staging environment in order
to test the new changes. Removes references to the Info object which is
no longer returned. Refactor `last_parsed` to be the last time we called
the API, since data is now on-demand, using an internal 6 hour cache.