Refactor hairstyles
Hairstyle images on the index page are now the sample images, instead of
loading the full spritesheet and manipulating it to only display the
first image. Samples are now organized into their own directory and the
API link has been updated to reference this. Increases the size of the
hairstyle spritesheet in case we eventually get unlockable
Viera/Hrothgar hairstyles.