Add character fetch jobs with Sidekiq
If a user visits a page with a stale character, a job is enqueued with
Sidekiq to fetch their latest data from XIVAPI. Consolidates character
fetch and refresh. Adds exception handling to character fetch since the
API will be throwing errors now when Lodestone is down, etc. Always
update the parse time on a fetch attempt to avoid spamming the API when
Lodestone is down. Removes recurring sync tasks & schedule.