Add character sync to API, remove /users
Character data is served right away, but they are added to the sync
queue if they are stale. This keeps the request from stalling for 8-20s.
API consumers can leverage the value of "last_parsed" to determine if
they are serving stale data to their users, and query later for fresh
data if necessary. Removes the /users endpoint since this would require
multiple character syncs and could quickly become unruly.