More character sync changes
Remove the manual refresh since there's no way to request a manual
update in the new API. Bump the "in queue" time up to 30 minutes to be
courteous to the new API, especially under maintenance. Fix an issue
when looking up untracked characters in the API so that it gives a
proper 404 error.