Respect character privacy for comparisons
Just like the selected character, we must also check the comparison
character's privacy whenever we look it up. Moves privacy verification
to the application controller since it should be checked basically
everywhere, not just on collections, but skip it for the API. Moves
Paper Trail before action to where it belongs.