Patch 5.1 fixes
Do not remove relations for Crafting sources, as this breaks the
automatic source update task. Properly check for existence and update
where appropriate for certain models, as they turned out not to be
static due to name changes. Update the Armoire load to check the item
description for gender since EquipRestriction currently isn't working.