Rework character verification
Characters can now be re-verified to different Discord accounts,
removing the issue of characters getting "stuck" on old/temporary
accounts. The verification is now done by the application directly
instead of via API for extra security. Users will now be alerted that
a character is not verified for their current user when browsing
manual collections, even if they are verified by another user.