Prepare for Patch 5.2 data
With the introduction of new Orchestrion categories, ID order is no
longer appropriate, so properly sort categories based on the in-game UI.
Updates the Orchestrion category load to allow for renaming of the old
categories. Fixes the source load to properly retrieve emotes now that
the commands are a list. Removes Moogle Treasure Trove from the navbar
since the event is no longer ongoing.