Disable PaperTrail individually on all load tasks
Individual tasks occasionally need to be re-run after an update, and can
fill the logs with unwanted changes. Instead of disabling PaperTrail at
the start of bulk loads, disable it within each individual task. Removes
unnecessary call to re-enable PaperTrail, since it is only disabled
within the scope of the task.