Improve Yo-kai tracking
Fixes an issue where the weapons collection was being given the
"minions" type. Adds Yo-kai weapons to the Medals table for clarity.
Moves the Yo-kai zones into the controller so it can be accessible for
this function. Rows in the Medals table are now selectable, allowing
users to highlight the weapon they are currently pursuing.