More character sync improvements
The check against the Sidekiq queue wasn't as responsive as we had hoped,
so restore `queued_at` specifically for tracking when the character was
queued for an automatic sync. DRYs up the sync check with a new `syncable?`
method. A character is syncable if their data is stale and they haven't
been queued for an automatic sync in the past 30 minutes.