Refactor navigation, upgrade to Font Awesome 5
Shifts navigation into a sticky sidebar, and retains locale/character
menus in a sticky topbar. The sidebar navigation will collapse on mobile
and be available via a hamburger menu in the topbar.

Overhauls column sizes and other styles on most of the pages
to properly fit content within the main canvas.

Upgrades to Font Awesome 5 in order to acquire sexy new icons and have
a wider selection of icons for sprucing up the navigation. Updates
references to old icon names as required.