Rework item unlocks, task cleanup
Item unlocks aren't always in Data[0], so create all of the collectables
first, then check all of the Data values for valid collectables. Sets
item_id for collectables in this same task. Collectables that depend on
unlocks for extra data now have that data set in a new item "extras"
task. Adds recipe data to items.