Minor fixes
Fixes formatting of spell descriptions since we lost newlines in the SC
refactor. Leverages the Verminion description formatter with a new
twist. Fixes an issue where spells with no sources were messing up the
table. Adds the proper description to the BLU Discord interaction for
spells and removes excess formatting.