Add Lodestone library
This library will take the place of XIVAPI for character lookups.
Fetches the character's profile, mounts, minions, and achievements
asynchronously.

In order to reduce the number of requests to the Lodestone, the full
list of achievements is only retrieved for new characters or
characters who have more than a full page worth of new achievements.