Improve event quest sources, link Armoire to Item
If we find an event quest source, create it iff the collectable has no
known sources. This will create placeholder quest links that are later
replaced with the event name, and won't re-generate the old quest link
after the fact.

Link items to Armoires when they are created. This ensures that quest
rewards can be properly linked to their collectables.