Add queued_at to Free Company
Allows users to refresh a free company even if all of its users are up
to date, as long as it hasn't been refreshed in the past 6 hours. This
is necessary in the event that every known character is up to date but
we still need to do a refresh to pull in new characters.