Fix private achievement shenanigans
Tracks achievement privacy with a proper boolean because we cannot and
should not be using a counter cache field like that. Removes characters
from the leaderboards if they set their achievements to private on the
Lodestone. Ensures characters who set their achievements to public have
their rankings restored without having to earn a new achievement.